pyproject.toml
[tool.poetry]
name = "uber"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.20.1"
pandas = "^1.4.2"
python-dotenv = "^0.20.0"
ipython = "^8.4.0"
notebook = "^6.4.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
